Extend exception handler decorator to Dataset; relocate decorator to utils
